# auth-service/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Pobeda Microservices API
  description: API для микросервисов базы отдыха «Победа», включающих управление пользователями, бронированиями и объектами (домиками и беседками).
  version: 1.0.0
servers:
  - url: https://auth-service-g23m.onrender.com
    description: Auth Service
  - url: https://booking-service-g1ea.onrender.com
    description: Booking Service
  - url: https://booking-admin-service.onrender.com
    description: Booking Admin Service
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Ошибка сервера
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Иван
        surname:
          type: string
          example: Иванов
        email:
          type: string
          format: email
          example: ivan@example.com
        phone:
          type: string
          example: +79123456789
        role:
          type: string
          enum: [user, admin]
          example: user
...
paths:
...
  /register:
    post:
      tags:
        - Auth_Service
      summary: Регистрация пользователя
      description: Создаёт нового пользователя с ролью "user". Пароль хешируется с использованием bcrypt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Иван
                surname:
                  type: string
                  example: Иванов
                email:
                  type: string
                  format: email
                  example: ivan@example.com
                phone:
                  type: string
                  example: +79123456789
                password:
                  type: string
                  example: password123
              required:
                - name
                - surname
                - email
                - phone
                - password
      responses:
        '201':
          description: Пользователь успешно зарегистрирован
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Пользователь зарегистрирован
        '400':
          description: Некорректные данные или email уже зарегистрирован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /login:
    post:
      tags:
        - Auth_Service
      summary: Вход пользователя
      description: Аутентифицирует пользователя и возвращает JWT-токен, действительный 7 дней.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: ivan@example.com
                password:
                  type: string
                  example: password123
              required:
                - email
                - password
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Email или пароль отсутствуют
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неверный email или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /profile:
    get:
      tags:
        - Auth_Service
      summary: Получение профиля пользователя
      description: Возвращает информацию о текущем пользователе на основе JWT-токена.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Auth_Service
      summary: Обновление профиля пользователя
      description: Обновляет данные пользователя (имя, фамилию, email, телефон). Email должен быть уникальным.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: Иван
                surname:
                  type: string
                  example: Иванов
                email:
                  type: string
                  format: email
                  example: ivan@example.com
                phone:
                  type: string
                  example: +79123456789
              required:
                - name
                - surname
                - email
                - phone
      responses:
        '200':
          description: Профиль успешно обновлён
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Профиль обновлён
        '400':
          description: Некорректные данные или email уже используется
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Неавторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  # Booking Admin Service
...